backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pandas
numpy
scikit-learn
beautifulsoup4
requests
pickle5
